# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.EzqTokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  handlers.CreateQueueRequest:
    properties:
      description:
        type: string
      eventTime:
        type: string
      location:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  handlers.LoginRequest:
    properties:
      provider:
        description: |-
          Auth provider
          example: google
          enums: google
        type: string
      token:
        description: |-
          Provider's access / ID token
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6...
        type: string
    type: object
  handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.Queue:
    properties:
      activeCashierIds:
        items:
          type: string
        type: array
      closureReason:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      eventTime:
        type: string
      id:
        type: string
      location:
        type: string
      mentorIds:
        items:
          type: string
        type: array
      name:
        type: string
      status:
        description: 'active, closed, ... // FIXME: enum?'
        type: string
    type: object
host: https://ezq.onrender.com
info:
  contact: {}
  description: Queue service API
  title: Easy Queue API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Refresh expired application's auth token
      parameters:
      - description: Refresh token value
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.EzqTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Refresh auth token
      tags:
      - auth
  /queues:
    post:
      consumes:
      - application/json
      description: Create a new queue
      parameters:
      - description: Queue payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateQueueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Queue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create queue
      tags:
      - queues
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"